Deep_Learning_with_PyTorch/requirements.txt
# On Ubuntu, install the following packages:
# - python3
# - python3-venv
#
# And then the virtual environment can be created by running:
#
# python3 -m venv .venv
# .venv/bin/pip install --upgrade pip
# .venv/bin/pip install -r requirements.txt
#
# Then use .venv/bin/python to run the various training scripts.

blosc
diskcache
h5py
jupyter
matplotlib
numpy
Pillow
scikit-image
scipy
SimpleITK
tensorflow

# The book was written targeting PyTorch 1.4, but everything in it should work fine with more up-to-date versions.
torch

# future is here due to a missing dep in caffe2
# future